requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
tenacity>=8.2.0
structlog>=24.1.0
pyyaml>=6.0.1
